Feature/block lanczos22

+++ Grid/util/Init.cc
@@ -167,13 +167,12 @@ void GridCmdOptionInt(std::string &str,int & val)
   return;
 }
 
-void GridCmdOptionFloat(std::string &str,float & val)
+void GridCmdOptionFloat(std::string &str,double & val)
 {
   std::stringstream ss(str);
   ss>>val;
   return;
 }
-
 
 void GridParseLayout(char **argv,int argc,
 		     Coordinate &latt_c,
